Adding `fileio` module with `OpenFile` class.
The `OpenFile` class is intended to provide a flexible way of
dealing with file streams using the `with` statement. It can
handle either path strings or already opened file streams as
the file argument, and can read/write gzipped and normal text
files.

The class still needs to be tested.