More work on `fileio` module.
Still untested and not used anywhere in the package outside the module.

Working on making `OpenFile` a more comprehensive "wrapper" class around
the `file` and `gzip.GzipFile` classes.

Also, adding `process_file_arg` function, which determines whether the
file argument is a string or not (then it assumes it is already an files
stream), and returns a tuple of the file stream and a boolean indicating
whether or not the function opened the file.  This can be useful for
code that can handle a file path string or a file stream argument. The
returned boolean can be then be used to determine whether or not the
file should be closed at that point in the code. I.e., if the file was
opened there, it should probably be closed. But, you may not want to
close a file that was opened by the caller.