Adding seq distance functions to `seqstats` module.
Adding `get_differences`, `distance`, and `per_site_distance` functions
to `seqstats` module. `get_differences` compares to sequences and
returns a dict mapping the indices of differences to a tuple of the
different character states. `distance` and `per_site_distance` use
`get_differences` and return the number of differences and the number of
differences divided by the sequence length, respectively.

Adding tests of all three functions to `test_seqstats.py`.