Updating `sample_iter` function.
Adding to additional arguments: `exclude` and `exclude_attribute`.  The
former is a list of elements to exclude when sampling from the
population, and the latter is the attribute to use to test whether an
element is in the exclusion list. The default for `exclude_attribute` is
`None`, in which case the exclude list is checked for the element
itself.

Adding additional test to `test_functions.py` that test the new
functionality of `sample_iter`.