Updating `align` module.
Adding `MafftAligner` and `MuscleAligner` classes. Renaming
the `align` function to `align_pair`, and adding new `align`
function, which performs multiple sequence alignments.

Adding `get_aligner` function, which handles the logic of looking for
alignment executables in PATH, and returning an instance of an aligner
class. This is called by `align` and `align_pair`.