Updating `get_differences` function.
Updating `get_differences` function of `seqstats` module. Rather than
reporting the sequence length of the compared sequences, it now returns
the number of sites used in calculating differences (i.e., when
`ignore_gaps = True` any sites with gaps are not counted).

As a result, when `get_distance` is called with `ignore_gaps = True`,
and `per_site = True`, it returns the proportion of sites that were
different out of the number of sites evaluated (rather than including
gaps in the denominator).

Updating tests in `test_seqstats.py` accordingly.