Updating `dataio` module.
Updating `BufferedIter` class and adding `seq_iter` and
`buffered_seq_iter` functions to `dataio` module.

Updating `BufferedIter` to allow objects to be appended to an instance.

`seq_iter` and `buffered_seq_iter` are similar to `get_seq_iter` and
`get_buffered_seq_iter` functions, but take a collection of file paths
or streams as the first argument (rather than a single file).