Updating `seq-file-slice.py` script.
Updating script for more efficient handling of sequential file formats.
The previous use of `SeqIO.write` for all file formats meant temporarily
storing `SeqRecord` objects in temp files via `BufferedIter`.

This is necessary for non-sequential file formats, but not for
sequential formats. If dealing with a sequential format, the script now
uses `write_seqs_to_files` function to write the sequences to files "on
the fly," which is much more efficient.