Updating `test_seqsum.py`.
Updating reverse-complement error tests involving mafft. In these tests,
an alignment of primates is used where the human sequence is
intentionally reverse-complemented. `seqsum.summarize_distances` is
called on these sequences to ensure the rev-comp error is caught.
However, for some versions of mafft, the tarsius sequence was also
divergent enough to cause a rev-comp error when `ignore_gaps` was
`True`. Changing to `ignore_gaps` to `False` which causes the desired
testing behavior.