Updating tests.
biopython no longer supports comparing SeqRecord instances for equality.
It used to be allowed and behave like the 'is' comparison (i.e., return
whether or not the SeqRecords were the same instance).

Changing several tests to use 'is' comparison to check confirm whether
to SeqRecords are the same instance or not, rather than using '==' (or
assertEqual/assertNotEqual).

Some entrez-related tests are still failing, but everything else is
passing.